+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -117,7 +117,7 @@ class ChromeProxyService implements VmServiceInterface {
   final _resumeAfterRestartEventsController =
       StreamController<String>.broadcast();
 
-  /// A global stream of resume events.
+  /// A global stream of resume events for hot restart.
   ///
   /// The values in the stream are the isolates IDs for the resume event.
   ///
@@ -126,6 +126,12 @@ class ChromeProxyService implements VmServiceInterface {
   /// subscriber to this stream.
   Stream<String> get resumeAfterRestartEventsStream =>
       _resumeAfterRestartEventsController.stream;
+
+  /// If non-null, a resume event should await the result of this after resuming
+  /// execution.
+  ///
+  /// This is used to complete a hot reload.
+  Future<void> Function()? _finishHotReloadOnResume;
 
   final _logger = Logger('ChromeProxyService');
 
@@ -225,6 +231,21 @@ class ChromeProxyService implements VmServiceInterface {
         'Received hot reload response but no pending completer was found (id: ${response.id})',
       );
     }
+  }
+
+  /// Reinitializes any caches so that they can be recomputed across hot reload.
+  // TODO(srujzs): We can maybe do better here than reinitializing all the data.
+  // Specifically, we can invalidate certain parts as we know what libraries
+  // will be stale, and therefore recompute information only for those libraries
+  // and possibly libraries that depend on them. Currently, there's no good
+  // separation between "existing" information and "new" information, making
+  // this difficult.
+  // https://github.com/dart-lang/webdev/issues/2628
+  Future<void> _reinitializeForHotReload() async {
+    final entrypoint = inspector.appConnection.request.entrypointPath;
+    await globalToolConfiguration.loadStrategy.trackEntrypoint(entrypoint);
+    _initializeEntrypoint(entrypoint);
+    await inspector.initialize();
   }
 
   /// Initializes metadata in [Locations], [Modules], and [ExpressionCompiler].
@@ -448,21 +469,15 @@ class ChromeProxyService implements VmServiceInterface {
   }
 
   /// Removes the breakpoints in the running isolate.
-  ///
-  /// [libraries] is a set of Dart libraries, where if non-null, only
-  /// breakpoints within those libraries are removed.
-  Future<void> disableBreakpoints({Set<String>? libraries}) async {
+  Future<void> disableBreakpoints() async {
     if (!_isIsolateRunning) return;
     final isolate = inspector.isolate;
 
-    for (final breakpoint in isolate.breakpoints?.toList() ?? <Breakpoint>[]) {
-      if (libraries == null ||
-          (breakpoint.location.script != null &&
-              // ignore: avoid-collection-methods-with-unrelated-types
-              libraries.contains(breakpoint.location.script.uri))) {
-        await (await debuggerFuture).removeBreakpoint(breakpoint.id!);
-      }
-    }
+    final debugger = await debuggerFuture;
+    await Future.wait([
+      for (final breakpoint in isolate.breakpoints ?? <Breakpoint>[])
+        debugger.removeBreakpoint(breakpoint.id!),
+    ]);
   }
 
   @override
@@ -1160,7 +1175,9 @@ class ChromeProxyService implements VmServiceInterface {
     String? rootLibUri,
     String? packagesUri,
   }) async {
-    _logger.info('Attempting a hot reload');
+    await isInitialized;
+    _checkIsolate('reloadSources', isolateId);
+
     ReloadReport getFailedReloadReport(String error) =>
         _ReloadReportWithMetadata(success: false)
           ..json = {
@@ -1172,7 +1189,7 @@ class ChromeProxyService implements VmServiceInterface {
       if (useWebSocket) {
         await _performWebSocketHotReload();
       } else {
-        await _performClientSideHotReload();
+        await _performClientSideHotReload(isolateId);
       }
     } catch (e) {
       _logger.info('Hot reload failed: $e');
@@ -1182,26 +1199,64 @@ class ChromeProxyService implements VmServiceInterface {
     return _ReloadReportWithMetadata(success: true);
   }
 
-  /// Performs a client-side hot reload by fetching libraries, disabling breakpoints, and invoking the reload.
-  Future<void> _performClientSideHotReload() async {
-    // Fetch the needed sources and libraries, disable breakpoints on the
-    // changed libraries, and then reload.
-    // TODO(srujzs): Re-map the breakpoints appropriately using events to
-    // trigger the client to re-register the breakpoints on the new sources.
-    // https://github.com/dart-lang/sdk/issues/60186
-    _logger.info('Issuing \$fetchLibrariesForHotReload request');
-    final librariesRemoteObject = await inspector.jsEvaluate(
-      '\$fetchLibrariesForHotReload();',
+  /// Performs a client-side hot reload by fetching libraries, handling
+  /// PausePostRequests, and invoking the reload.
+  Future<void> _performClientSideHotReload(String isolateId) async {
+    _logger.info('Attempting a hot reload');
+
+    // Initiate a hot reload.
+    _logger.info('Issuing \$dartHotReloadStartDwds request');
+    await inspector.jsEvaluate(
+      '\$dartHotReloadStartDwds();',
       awaitPromise: true,
       returnByValue: true,
     );
-    _logger.info('\$fetchLibrariesForHotReload request complete.');
-    final libraries =
-        (librariesRemoteObject.value as List<dynamic>).toSet().cast<String>();
-    await disableBreakpoints(libraries: libraries);
-    _logger.info('Issuing \$dartHotReloadDwds request');
-    await inspector.jsEvaluate('\$dartHotReloadDwds();', awaitPromise: true);
-    _logger.info('\$dartHotReloadDwds request complete.');
+
+    if (!pauseIsolatesOnStart) {
+      // Finish hot reload immediately.
+      _logger.info('Issuing \$dartHotReloadEndDwds request');
+      await inspector.jsEvaluate(
+        '\$dartHotReloadEndDwds();',
+        awaitPromise: true,
+      );
+      _logger.info('\$dartHotReloadEndDwds request complete.');
+      // TODO(srujzs): Supposedly Dart DevTools uses a kIsolateReload event
+      // for breakpoints? We should confirm and add tests before sending the
+      // event.
+      return;
+    }
+    // If `pause_isolates_on_start` is enabled, pause and then the reload
+    // should finish later after the client removes breakpoints, reregisters
+    // breakpoints, and resumes.
+    _finishHotReloadOnResume = () async {
+      // Client finished setting breakpoints, called resume, and now the
+      // execution has resumed. Finish the hot reload so we start executing
+      // the new code instead.
+      _logger.info('Issuing \$dartHotReloadEndDwds request');
+      await inspector.jsEvaluate(
+        '\$dartHotReloadEndDwds();',
+        awaitPromise: true,
+      );
+      _logger.info('\$dartHotReloadEndDwds request complete.');
+    };
+
+    // Pause and wait for the pause to occur before managing breakpoints.
+    final pausedEvent = _firstStreamEvent('Debug', EventKind.kPauseInterrupted);
+    await pause(isolateId);
+    await pausedEvent;
+
+    await _reinitializeForHotReload();
+
+    // This lets the client know that we're ready for breakpoint management
+    // and a resume.
+    _streamNotify(
+      'Debug',
+      Event(
+        kind: EventKind.kPausePostRequest,
+        timestamp: DateTime.now().millisecondsSinceEpoch,
+        isolate: inspector.isolateRef,
+      ),
+    );
   }
 
   /// Performs a WebSocket-based hot reload by sending a request and waiting for a response.
@@ -1269,20 +1324,24 @@ class ChromeProxyService implements VmServiceInterface {
       _resumeAfterRestartEventsController.add(isolateId);
       return Success();
     }
+
     if (inspector.appConnection.isStarted) {
-      return captureElapsedTime(() async {
+      await captureElapsedTime(() async {
         await isInitialized;
         await isStarted;
         _checkIsolate('resume', isolateId);
-        return await (await debuggerFuture).resume(
-          step: step,
-          frameIndex: frameIndex,
-        );
+        final debugger = await debuggerFuture;
+        return await debugger.resume(step: step, frameIndex: frameIndex);
       }, (result) => DwdsEvent.resume(step));
     } else {
       inspector.appConnection.runMain();
-      return Success();
     }
+    // Finish the hot reload if needed.
+    if (_finishHotReloadOnResume != null) {
+      await _finishHotReloadOnResume!();
+      _finishHotReloadOnResume = null;
+    }
+    return Success();
   }
 
   /// This method is deprecated in vm_service package.
@@ -1573,6 +1632,11 @@ class ChromeProxyService implements VmServiceInterface {
     final controller = _streamControllers[streamId];
     if (controller == null) return;
     controller.add(event);
+  }
+
+  Future<void> _firstStreamEvent(String streamId, String eventKind) {
+    final controller = _streamControllers[streamId]!;
+    return controller.stream.firstWhere((event) => event.kind == eventKind);
   }
 
   Future<void> _handleDeveloperLog(