+++ dwds/test/frontend_server_hot_restart_correctness_test.dart
@@ -16,7 +16,7 @@ import 'fixtures/context.dart';
 
 void main() {
   // Enable verbose logging for debugging.
-  final debug = false;
+  const debug = false;
   final canaryFeatures = true;
   final moduleFormat = ModuleFormat.ddc;
   final compilationMode = CompilationMode.frontendServer;