+++ dwds/test/hot_reload_test.dart
@@ -22,7 +22,7 @@ const newString = 'Bonjour le monde!';
 
 void main() {
   // Enable verbose logging for debugging.
-  final debug = false;
+  const debug = false;
   final provider = TestSdkConfigurationProvider(
     verbose: debug,
     canaryFeatures: true,