+++ dwds/web/reloader/restarter.dart
@@ -9,13 +9,13 @@ abstract class Restarter {
   /// not.
   Future<bool> restart({String? runId, Future? readyToRunMain});
 
-  /// Performs a hot reload using the sources and libraries computes in the
-  /// previous call to [fetchLibrariesForHotReload].
-  Future<void> reload();
+  /// After a previous call to [hotReloadStart], completes the hot
+  /// reload by pushing the libraries into the Dart runtime.
+  Future<void> hotReloadEnd();
 
-  /// Computes the sources and libraries to reload and returns the list of
-  /// libraries using [hotReloadSourcesPath] as the path to a JSONified list of
-  /// maps which follows the following format:
+  /// Computes the sources and libraries to reload, loads them into the page,
+  /// and returns the list of libraries using [hotReloadSourcesPath] as the path
+  /// to a JSONified list of maps which follows the following format:
   ///
   /// ```json
   /// [
@@ -30,7 +30,5 @@ abstract class Restarter {
   /// bundle.
   /// `libraries`: An array of strings containing the libraries that were
   /// compiled in `src`.
-  Future<JSArray<JSString>> fetchLibrariesForHotReload(
-    String hotReloadSourcesPath,
-  );
+  Future<JSArray<JSString>> hotReloadStart(String hotReloadSourcesPath);
 }