+++ dwds/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added WebSocket-based hot reload support: `reloadSources` in `ChromeProxyService` and `DevHandler` now handle hot reload requests and responses over WebSockets.
 - Refactored the injected client to use a reusable function for handling hot reload requests and responses over WebSockets.
-- Added support for breakpoint registering on a hot restart with the DDC library bundle format using PausePostRequests.
+- Added support for breakpoint registering on a hot restart and on a hot reload with the DDC library bundle format using PausePostRequests.
 
 ## 24.3.10